put parse functions into utils
